Auth + dashboard wiring, SSE docs, and Nginx compose integration
- server.js: add routes for /login, /register, /dashboard
- routes/user.js: fallback to personal_newAccount when getnewaddress unavailable
- public/exchange.html: add quick links to auth pages
- README_EXCHANGE_LISTING.md: document SSE, auth/user endpoints, account pages
- compose.yaml: add JWT_SECRET, persistent backend_data volume, and nginx service
- nginx.conf: proxy auth pages; simplify HTTP-only until SSL certs mounted; backend upstream

Deployment notes: backend rebuilt and auth verified; compose-managed nginx pending SSL provisioning.

+++ src/main.cpp
@@ -39,7 +39,7 @@ json handleSendToAddress(const json &jreq) {
         return {{"error", "Invalid parameters"}};
     }
     std::string to = jreq["params"][0].get<std::string>();
-    int amount = jreq["params"][1].get<int>();
+    long long amount = jreq["params"][1].get<long long>();
     return {{"result", db->sendToAddress(to, amount)}};
 }
